+++ src/main/python/game/scenes/ExitScene.py
@@ -81,7 +81,8 @@ class ExitScene(Scene):
 
     def loop(self, tick):
         dt = tick / 1000
-        self.game_data.background.loop(dt)
+
+        self.game_data.background.loop(dt, iterate_offset=True)
 
         # Handle events
         for event in pygame.event.get():