+++ src/main/python/game/scenes/HighscoreScene.py
@@ -206,7 +206,7 @@ class HighscoreScene(Scene):
     def loop(self, tick):
         dt = tick / 1000
 
-        self.game_data.background.loop(dt)
+        self.game_data.background.loop(dt, iterate_offset=True)
 
         # Handle events
         for event in pygame.event.get():