+++ src/main/python/game/scenes/MenuScene.py
@@ -298,7 +298,7 @@ class MenuScene(Scene):
     def loop(self, tick):
         dt = tick / 1000
 
-        self.game_data.background.loop(dt)
+        self.game_data.background.loop(dt, iterate_offset=True)
 
         # Handle events
         for event in pygame.event.get():